blit_open_gl_texture_to_map_position is really supposed to take floats not doubles

+++ src/proto.h
@@ -220,7 +220,7 @@ EXTERN void GL_HighlightRectangle ( SDL_Surface* Surface , SDL_Rect Area , unsig
 EXTERN void ShowInventoryScreenBackground ( void );
 EXTERN void show_character_screen_background ( void );
 EXTERN int safely_initialize_our_default_open_gl_parameters ( void ) ;
-EXTERN void blit_open_gl_texture_to_map_position ( iso_image * our_floor_iso_image , float our_col , float our_line , double r , double g , double b , int highlight_texture, int blend ) ;
+EXTERN void blit_open_gl_texture_to_map_position ( iso_image * our_floor_iso_image , float our_col , float our_line , float r , float g , float b , int highlight_texture, int blend ) ;
 EXTERN void blit_zoomed_open_gl_texture_to_map_position ( iso_image * our_floor_iso_image , float our_col , float our_line , float r , float g , float b , int highlight_texture, int blend ) ;
 EXTERN void blit_open_gl_texture_to_screen_position ( iso_image * our_floor_iso_image , int x , int y , int set_gl_parameters ) ;
 EXTERN void blit_semitransparent_open_gl_texture_to_screen_position ( iso_image * our_floor_iso_image , int x , int y , float scale_factor ) ;
